Restriction of VR sliders (slits/slitwidth), amplitude after plate is a bit lower now
