+++ unity/Assets/0Refactored/scenes/experiments/HuygensPrinciple/HuygensPrinciple.pc.unity
@@ -202,7 +202,7 @@ PrefabInstance:
     - target: {fileID: 602218157930835559, guid: 40aab8bc34f575e4eaddf6b9545c0eef,
         type: 3}
       propertyPath: m_LocalPosition.y
-      value: 7.19599
+      value: 7.145
       objectReference: {fileID: 0}
     - target: {fileID: 602218157930835559, guid: 40aab8bc34f575e4eaddf6b9545c0eef,
         type: 3}
@@ -2106,6 +2106,11 @@ PrefabInstance:
       propertyPath: planeMesh
       value: 
       objectReference: {fileID: 1882232453}
+    - target: {fileID: 1846387199857951263, guid: 0e21da698c6f373409b2eb6121cdc3a1,
+        type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 1.127
+      objectReference: {fileID: 0}
     - target: {fileID: 1846387200166215732, guid: 0e21da698c6f373409b2eb6121cdc3a1,
         type: 3}
       propertyPath: m_LocalRotation.x
