+++ unity/Assets/0Refactored/scenes/experiments/HuygensPrinciple/scripts/RestrictSliderValuesVR.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 1ac4a879fc7648b5bc89a1018653ffa2
+timeCreated: 1589259840
